+++ src/main/routes/routes.ts
@@ -88,21 +88,6 @@ export default function(app: Application): void {
     });
   }
 
-  function forgotPasswordRedirect(req, res, next): void {
-    const body = JSON.stringify(req.body);
-    if (body.includes('AADB2C90118')) {
-      const CLIENT_ID = config.get('secrets.pip-ss-kv.CLIENT_ID');
-      const B2C_URL = config.get('secrets.pip-ss-kv.B2C_URL');
-      const REDIRECT_URL = `${FRONTEND_URL}/password-change-confirmation`;
-      const POLICY_URL = `${B2C_URL}/oauth2/v2.0/authorize?p=${authenticationConfig.FORGOT_PASSWORD_POLICY}` +
-        `&client_id=${CLIENT_ID}&nonce=defaultNonce&redirect_uri=${REDIRECT_URL}` +
-        '&scope=openid&response_type=id_token&prompt=login';
-      res.redirect(POLICY_URL);
-      return;
-    }
-    return next();
-  }
-
   // Public paths
   app.get('/*', globalAuthGiver);
   app.post('/*', globalAuthGiver);
@@ -121,7 +106,7 @@ export default function(app: Application): void {
   app.get('/hearing-list', app.locals.container.cradle.hearingListController.get);
   app.get('/password-change-confirmation', app.locals.container.cradle.passwordChangeController.get);
   app.get('/login', passport.authenticate(authType, { failureRedirect: '/'}), regenerateSession);
-  app.post('/login/return', forgotPasswordRedirect, passport.authenticate(authType, { failureRedirect: '/view-option'}),
+  app.post('/login/return', adminAuthentication.forgotPasswordRedirect, passport.authenticate(authType, { failureRedirect: '/view-option'}),
     (_req, res) => {adminAuthentication.isAdminUser(_req) ? res.redirect('/admin-dashboard') : res.redirect('/account-home');});
   app.get('/logout', (_req, res) => {adminAuthentication.isAdminUser(_req) ?
     logOut(_req, res, `${FRONTEND_URL}/login?p=`+ authenticationConfig.ADMIN_POLICY) : logOut(_req, res, `${FRONTEND_URL}/view-option`);});
